Make search button invisible if no selected currencies

+++ app/src/main/java/ua/nanit/extop/monitoring/CommonCalculator.kt
@@ -4,7 +4,7 @@ import ua.nanit.extop.monitoring.data.Rate
 
 class CommonCalculator : RateCalculator {
 
-    override fun calculate(rates: List<Rate>, dir: Direction, amount: Float, ) {
+    override fun calculate(rates: List<Rate>, dir: Direction, amount: Double) {
         when (dir) {
             Direction.IN -> {
                 calcIn(rates, amount)
@@ -15,7 +15,7 @@ class CommonCalculator : RateCalculator {
         }
     }
 
-    private fun calcIn(rates: List<Rate>, amount: Float) {
+    private fun calcIn(rates: List<Rate>, amount: Double) {
         for (rate in rates) {
             val course = rate.amountOut / rate.amountIn
 
@@ -25,7 +25,7 @@ class CommonCalculator : RateCalculator {
         }
     }
 
-    private fun calcOut(rates: List<Rate>, amount: Float) {
+    private fun calcOut(rates: List<Rate>, amount: Double) {
         for (rate in rates) {
             val course = rate.amountIn / rate.amountOut
