+++ app/src/main/java/ua/nanit/extop/monitoring/RateCalculator.kt
@@ -4,6 +4,6 @@ import ua.nanit.extop.monitoring.data.Rate
 
 interface RateCalculator {
 
-    fun calculate(rates: List<Rate>, dir: Direction, amount: Float)
+    fun calculate(rates: List<Rate>, dir: Direction, amount: Double)
 
 }
